Update API response formats and improve search functionality
- Standardize /api/words and /api/search responses to use BaseListResp format
- Update default page size from 24 to 12 for better mobile experience
- Enhance search functionality to return total count with results
- Update vocabulary data structures to include complete field information
- Add BaseListResp type definition for consistent API responses
- Update API documentation to reflect all changes
- Improve error handling and logging in search and word operations

+++ internal/dto/vocabulary_dto.go
@@ -12,7 +12,7 @@ type VocabularyList struct {
 // VocabularyPage represents a paginated result of vocabulary words
 type VocabularyPage struct {
 	Words      []table.Word `json:"words"`
-	TotalCount int          `json:"totalCount"`
+	TotalCount int64        `json:"totalCount"`
 	PageNumber int          `json:"pageNumber"`
 	PageSize   int          `json:"pageSize"`
 	TotalPages int          `json:"totalPages"`
@@ -20,17 +20,17 @@ type VocabularyPage struct {
 
 // VocabularyStats represents vocabulary statistics
 type VocabularyStats struct {
-	TotalWords   int                    `json:"totalWords"`
-	TotalPages   int                    `json:"totalPages"`
+	TotalWords  int                    `json:"totalWords"`
+	TotalPages  int                    `json:"totalPages"`
 	WordsByPage map[string]interface{} `json:"wordsByPage"`
 }
 
 // WordWithMarkStatus represents a word with its mark status for a user
 type WordWithMarkStatus struct {
 	table.Word
-	IsMarked   bool `json:"isMarked"`
-	MarkCount  int  `json:"markCount"`
-	MarkedAt   int64 `json:"markedAt,omitempty"`
+	IsMarked  bool  `json:"isMarked"`
+	MarkCount int   `json:"markCount"`
+	MarkedAt  int64 `json:"markedAt,omitempty"`
 }
 
 // VocabularyPageWithMarks represents a paginated result with mark status
