+++ internal/dto/vocabulary_legacy.go
@@ -12,7 +12,7 @@ type WordList struct {
 // Page represents a paginated result of vocabulary words (legacy structure for compatibility)
 type Page struct {
 	Words      []table.Word `json:"words"`
-	TotalCount int          `json:"total_count"`
+	TotalCount int64        `json:"total_count"`
 	PageNumber int          `json:"page_number"`
 	PageSize   int          `json:"page_size"`
 	TotalPages int          `json:"total_pages"`
